Mejoras de service

+++ ms-personas/src/main/java/com/banco/digital/ms_personas/model/UserState.java
@@ -14,7 +14,7 @@ public class UserState {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idestado_usuario")
-    private Long idUserState;
+    private Integer idUserState;
 
     @Column(name = "descripcion")
     private String description;
